+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionCommandExecutor.java
@@ -1,4 +1,4 @@
-package com.techcourse.service.transaction;
+package org.springframework.transaction.support;
 
 @FunctionalInterface
 public interface TransactionCommandExecutor {
