+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionQueryExecutor.java
@@ -1,4 +1,4 @@
-package com.techcourse.service.transaction;
+package org.springframework.transaction.support;
 
 @FunctionalInterface
 public interface TransactionQueryExecutor<T> {
