+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionTemplate.java
@@ -1,9 +1,8 @@
-package com.techcourse.service.transaction;
+package org.springframework.transaction.support;
 
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.datasource.DataSourceUtils;
 import org.springframework.transaction.TransactionException;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
